+++ src/index.ts
@@ -15,6 +15,7 @@ export * from "./extraction/infoTable";
 export * from "./extraction/thirteenFValueUnits";
 export * from "./extraction/cusipTickerCrosswalk";
 export * from "./extraction/resolveHoldingTickers";
+export * from "./extraction/issuerTicker";
 export * from "./extraction/tsv";
 export * from "./lake/holdingsTables";
 export * from "./lake/holdingsShardRows";